requirements.txt
appdirs==1.4.3
asgiref==3.4.1
astroid==2.8.3
attrs==19.3.0
autopep8==1.6.0
beautifulsoup4==4.8.2
black==21.9b0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
colorama==0.4.3
configobj==5.0.6
cryptography==2.8
defusedxml==0.7.1
distro==1.4.0
dj-database-url==0.5.0
django-admin-interface==0.17.3
django-bootstrap4==3.0.1
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-grappelli==2.15.3
django-jet==1.0.8
django-js-asset==1.2.2
django-robots==4.0
djangorestframework==3.12.4
entrypoints==0.3
gunicorn==20.1.0
httplib2==0.14.0
humanize==1.0.0
idna==2.8
ifaddr==0.1.6
IMDbPY==6.8
isort==5.9.3
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
loguru==0.5.3
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
netaddr==0.7.19
netifaces==0.10.4
numpy==1.21.3
oauthlib==3.1.0
packaging==20.3
pathspec==0.9.0
pexpect==4.6.0
pgspecial==1.9.0
platformdirs==2.4.0
prompt-toolkit==2.0.10
protobuf==3.6.1
psutil==5.5.1
psycopg2-binary==2.9.1
pycairo==1.16.2
pycodestyle==2.8.0
pycrypto==2.6.1
Pygments==2.3.1
pyinotify==0.9.6
PyJWT==2.3.0
pylint==2.11.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
python-xlib==0.23
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2021.10.23
reportlab==3.5.34
requests==2.22.0
requests-file==1.4.3
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==1.9.5
sqlparse==0.4.2
tabulate==0.8.6
terminaltables==3.1.0
tinycss2==1.0.2
tldextract==2.2.1
toml==0.10.0
tomli==1.2.1
typing-extensions==3.10.0.2
Unidecode==1.1.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.13.2
